+++ crates/da-clients/ethereum/src/config.rs
@@ -1,4 +1,10 @@
+use crate::EthereumDaClient;
+use alloy::network::Ethereum;
+use alloy::providers::ProviderBuilder;
+use alloy::rpc::client::RpcClient;
 use serde::{Deserialize, Serialize};
+use std::str::FromStr;
+use url::Url;
 use utils::settings::Settings;
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -15,5 +21,13 @@ impl EthereumDaConfig {
             memory_pages_contract: settings.get_settings_or_panic("MEMORY_PAGES_CONTRACT_ADDRESS"),
             private_key: settings.get_settings_or_panic("PRIVATE_KEY"),
         })
+    }
+
+    pub async fn build_client(&self) -> EthereumDaClient {
+        let client =
+            RpcClient::new_http(Url::from_str(self.rpc_url.as_str()).expect("Failed to parse SETTLEMENT_RPC_URL"));
+        let provider = ProviderBuilder::<_, Ethereum>::new().on_client(client);
+
+        EthereumDaClient { provider }
     }
 }