+++ crates/settlement-clients/ethereum/src/tests/mod.rs
@@ -240,7 +240,7 @@ mod settlement_client_tests {
         // Asserting, Expected to receive transaction hash.
         assert!(!update_state_result.is_empty(), "No transaction Hash received.");
 
-        sleep(Duration::from_secs(5)).await;
+        sleep(Duration::from_secs(10)).await;
         ethereum_settlement_client
             .wait_for_tx_finality(update_state_result.as_str())
             .await